Blazzing Pizza Done

+++ MyBlazorHybridApp/MyBlazorHybridApp.Shared/Model/Order.cs
@@ -16,7 +16,7 @@ namespace MyBlazorHybridApp.Shared.Model
 
         public List<Pizza> Pizzas { get; set; } = new List<Pizza>();
 
-        public decimal GetTotalPrice() => Pizzas.Sum(p => p.GetTotalPrice());
+        public double GetTotalPrice() => Pizzas.Sum(p => p.GetTotalPrice());
 
         public string GetFormattedTotalPrice() => GetTotalPrice().ToString("0.00");
     }
